+++ src/main/java/com/ludo/study/studymatchingplatform/auth/service/google/GoogleProfileRequestService.java
@@ -6,7 +6,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
-import com.ludo.study.studymatchingplatform.auth.naver.repository.InMemoryClientRegistrationAndProviderRepository;
+import com.ludo.study.studymatchingplatform.auth.repository.InMemoryClientRegistrationAndProviderRepository;
 import com.ludo.study.studymatchingplatform.auth.service.google.vo.GoogleUserInfo;
 import com.ludo.study.studymatchingplatform.user.domain.Social;
